requirements.txt
pytest==7.4.4
requests==2.31.0
allure-pytest==2.13.5
allure-python-commons==2.13.5